Crate pallet_voting

Voting pallet

The voting pallet provides methods to manage the voting processing through house council vote and investors voting

Overview

This pallet manage the voting of a proposal by the House Council and an investor assemblee

Dispatchable Functions

‘submit_proposal’ - an account with the seller role submit a proposal for a house purchase

‘call_democracy_proposal’ - configure a proposal to go through the democracy vote processing

‘call_dispatch’ - execute the house purchase proposal

‘council_vote’ - a member of the House Council vote for the first step going through the Collective pallet

‘council_close_vote’ - a member of the House Council close the collective vote session

‘investor_vote’ - an investor vote for the proposal during the democracy voting step

Re-exports

pub use pallet_collective as COLL;

pub use pallet_democracy as DEMO;

pub use pallet_roles as ROLES;

pub use weights::WeightInfo;

pub use pallet::*;

pub use super::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_template

Structs

VotingProposal

Traits

Currency

Abstraction over a fungible assets system.

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

Type Definitions

AccountIdOf

BalanceOf

BlockNumberOf

Derive Macros

Decode

Derive parity_scale_codec::Decode and for struct and enum.

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.

RuntimeDebug