Trait pallet_voting::pallet::Config

pub trait Config: Config + Config<Instance1> + Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Call: Parameter + UnfilteredDispatchable<Origin = <Self as Config>::Origin> + From<Call<Self>> + GetDispatchInfo;
    type WeightInfo: WeightInfo;
    type Delay: Get<Self::BlockNumber>;
    type CheckDelay: Get<Self::BlockNumber>;
    type InvestorVoteAmount: Get<u128>;
    type LocalCurrency: ReservableCurrency<Self::AccountId>;
    type HouseCouncilOrigin: EnsureOrigin<<Self as Config>::Origin>;
    type CheckPeriod: Get<Self::BlockNumber>;
    type MinimumDepositVote: Get<BalanceOf<Self>>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Call: Parameter + UnfilteredDispatchable<Origin = <Self as Config>::Origin> + From<Call<Self>> + GetDispatchInfo

type WeightInfo: WeightInfo

type Delay: Get<Self::BlockNumber>

type CheckDelay: Get<Self::BlockNumber>

type InvestorVoteAmount: Get<u128>

type LocalCurrency: ReservableCurrency<Self::AccountId>

type HouseCouncilOrigin: EnsureOrigin<<Self as Config>::Origin>

type CheckPeriod: Get<Self::BlockNumber>

type MinimumDepositVote: Get<BalanceOf<Self>>